requirements.txt
telethon
selenium
beautifulsoup4
jdatetime
psycopg2-binary
sqlalchemy